slywriter-ui/backend/requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
alembic==1.12.1

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
google-auth==2.23.4

# AI Integration
openai==1.6.1

# WebSocket & Async
websockets==12.0
aiohttp==3.9.1

# Typing Automation
pyautogui==0.9.54
keyboard==0.13.5
pyperclip==1.8.2

# Utils
pydantic==2.5.2
python-dateutil==2.8.2
requests==2.31.0

# Voice & Audio (Optional)
SpeechRecognition==3.10.1
pyttsx3==2.90

# Development
pytest==7.4.3
pytest-asyncio==0.21.1

# CORS and Security
fastapi-cors==0.0.6